pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "fastapi-base"
version = "0.1.0"
description = "FastAPI application skeleton (async SQLAlchemy 2, Alembic)"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
